Move Qwen trace utility helpers out of main
